+++ packages/grid/src/config/TableConfig.ts
@@ -1,6 +1,7 @@
 export const tableBorderColor = "var(--lhds-color-ui-300)";
 export const tableBorderColorExpanded = "var(--lhds-color-blue-500)";
 export const tableBackgroundColorExpanded = "var(--lhds-color-blue-50)";
+export const tableBackgroundHoverColorExpanded = "var(--lhds-color-blue-100)";
 export const tableBorder = `1px solid ${tableBorderColor}`;
 export const tableBorderLeft = `2px solid transparent`;
 export const tableBorderLeftExpanded = `2px solid ${tableBorderColorExpanded}`;