+++ CHANGELOG.md
@@ -30,10 +30,10 @@ It returns the same actions, but wrapped with reducerIdGate action.
 ### StandardTable
 
 - Sort order icon `amount` variant now has inverted arrow direction.
-
 - When using editable cells, `editorValue` was not always reset and could
 become inconsistent with the value of the data.
 `editorValue` is now always reset when editor for a cell is opened.
+- `rowBackgroundResolver` can now return color or object containing `background` and `hoverBackground` fields.
 
 ### Dependencies